weatherapp/requirements.txt
amqp==2.3.2
Babel==2.6.0
billiard==3.5.0.3
celery==4.1.1
certifi==2018.4.16
chardet==3.0.4
Django==2.0.5
django-cors-headers==2.2.0
djangorestframework==3.8.2
flower==0.9.2
idna==2.6
kombu==4.2.1
pkg-resources==0.0.0
pytz==2018.4
requests==2.18.4
tornado==5.0.2
urllib3==1.22
vine==1.1.4
